remove react-composites files that are no longer used. move one to one examples to storybook project
