--- packages/react-composites/src/composites/common/index.ts
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-export * from './ErrorBar';
