+++ packages/react-composites/src/composites/index.ts
@@ -4,7 +4,3 @@
 export * from './ChatComposite';
 export * from './CallComposite';
 export * from './MeetingComposite';
-export { ErrorBar } from './common';
-
-// Export for storybook, these being exported should be re-evaluated when composites are pure under the new architecture.
-export * from './IncomingCallAlerts';
