+++ packages/react-composites/src/index.ts
@@ -1,13 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export * from './composites';
-export type { CommunicationUiErrorArgs, CommunicationUiErrorInfo, CommunicationUiErrorSeverity } from './types';
-
-export { propagateError, createAzureCommunicationUserCredential } from './utils';
-export {
-  CommunicationUiError,
-  CommunicationUiErrorCode,
-  CommunicationUiErrorFromError
-} from './types/CommunicationUiError';
-export type { DevicePermissionState } from './types/DevicePermission';
+export * from './composites/ChatComposite';
+export * from './composites/CallComposite';
+export * from './composites/MeetingComposite';
