--- packages/react-composites/src/types/DevicePermission.ts
@@ -1,5 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-export type DevicePermissionState = 'Granted' | 'Denied' | 'Unknown';
-export type DevicePermissionType = 'Camera' | 'Microphone';
