--- packages/react-composites/src/types/ParticipantStream.ts
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import { RemoteParticipant, RemoteVideoStream } from '@azure/communication-calling';
-
-export type ParticipantStream = {
-  user: RemoteParticipant;
-  stream: RemoteVideoStream | undefined;
-};
