--- packages/react-composites/src/types/index.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-export * from './CommunicationUiError';
-export * from './DevicePermission';
-export * from './CommunicationUiError';
