+++ packages/storybook/stories/Examples/IncomingCallAlerts/IncomingCallToast.tsx
@@ -2,9 +2,9 @@
 // Licensed under the MIT license.
 
 import { Stack } from '@fluentui/react';
-import { IncomingCallToast as IncomingCallToastComponent } from '@internal/react-composites';
 import { text, files } from '@storybook/addon-knobs';
 import React from 'react';
+import { IncomingCallToast as IncomingCallToastComponent } from './components';
 
 export const IncomingCallToast: () => JSX.Element = () => {
   const callerName = text('Caller Name', 'John Doe');
